Trx Earn Bot/bot/requirements.txt
python-telegram-bot
firebase-admin
python-dotenv
firebase-admin>=6.0.0
